+++ prototype/tests/conftest.py
@@ -1,35 +1,6 @@
 import pytest
-import os
-import sys
-
-
-def _set_source_root_parent(source_root_parent):
-    """
-    add everything under `source_root_parent` to the list of source roots
-    e.g. if you call this with param 'submodules'
-    and you have the file structure
-
-    project/
-        submodules/
-            foo-src/
-                foo/
-            bar-src/
-                bar/
-
-    (where foo and bar are python modules)
-    then foo and bar would become top-level importable
-
-    """
-    filedir = os.path.dirname(__file__)
-    submodules_list = os.listdir(os.path.join(filedir, source_root_parent))
-    for d in submodules_list:
-        if d == "__init__.py" or d == '.' or d == '..':
-            continue
-        sys.path.insert(1, os.path.join(filedir, source_root_parent, d))
-
-    sys.path.append(os.path.join(filedir, source_root_parent))
-
-_set_source_root_parent('../submodules')
+from patch_path import patch_path
+patch_path()
 
 from mobile_endpoint import create_app
 from mobile_endpoint.models import db
